# mental-klinik-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem informasi klinik kesehatan mental",
        "title": "Klinik API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/assessments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated list of assessments, with optional filter by patient ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Get all assessments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by patient ID",
                        "name": "patientId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedAssessmentsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat data assessment baru untuk pasien tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Buat assessment baru",
                "parameters": [
                    {
                        "description": "Data assessment baru",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAssessmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAssessmentSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/assessments/patient/{patientId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua data assessment yang dimiliki oleh pasien tertentu, termasuk hasil prediksi jika tersedia",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Mendapatkan semua assessment berdasarkan ID pasien",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AssessmentResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/assessments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail satu assessment berdasarkan ID, termasuk data pasien (mini) dan hasil prediksi jika ada",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Mendapatkan assessment berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Assessment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAssessmentByIDSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui tanggal dan jawaban assessment, serta memperbarui waktu terakhir diperbarui (updatedAt)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Memperbarui assessment berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Assessment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data assessment yang diperbarui",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssessmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAssessmentSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus assessment dari database berdasarkan ID yang diberikan",
                "tags": [
                    "Assessments"
                ],
                "summary": "Menghapus assessment berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Assessment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageDeleteAssesmentResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AssessmentResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "patient": {
                    "description": "dari relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.PatientMiniResponse"
                        }
                    ]
                },
                "patientId": {
                    "type": "string"
                },
                "prediction": {
                    "description": "nullable",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.PredictionResponse"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.AssessmentUpdateResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "date": {
                    "type": "string",
                    "example": "2025-07-17T15:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "d72f8239-09e7-4b23-a897-3ef6211f6b0a"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-07-17T15:05:30Z"
                }
            }
        },
...
        "dto.CreateAssessmentRequest": {
            "type": "object",
            "required": [
                "date",
                "patientId"
            ],
            "properties": {
                "answers": {
                    "description": "Raw JSON",
                    "type": "object",
                    "additionalProperties": true
                },
                "date": {
                    "description": "sesuai format date ISO",
                    "type": "string"
                },
                "patientId": {
                    "type": "string"
                }
            }
        },
        "dto.CreateAssessmentSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AssessmentResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
        "dto.GetAssessmentByIDSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AssessmentResponse"
                }
            }
        },
...
        "dto.MessageDeleteAssesmentResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginatedAssessmentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AssessmentResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PatientMiniResponse": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "dto.PredictionResponse": {
            "type": "object",
            "properties": {
                "assessmentId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "probabilityScore": {
                    "type": "number"
                },
                "resultLabel": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateAssessmentRequest": {
            "type": "object",
            "required": [
                "date"
            ],
            "properties": {
                "answers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateAssessmentSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AssessmentUpdateResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Assessment berhasil diperbarui"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}